pyproject.toml
[project]
name = "academic-helper"
version = "0.1.0"
description = "Academic research assistant — literature search, gap discovery, content evaluation"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "pyyaml>=6.0",
]

[project.scripts]
academic-helper = "academic_helper.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/academic_helper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
